jp - add bootstrap 4
Added bootstrap 4 to the Gemfile as well as another gem that lets tooltips
position themselves nicely upon hover. Also required/included bootstrap in
application.js/scss files so that we can use them anywhere in the app. The
application.html.erb file includes everything from application by default, so it
will be pretty nice as we don't have to clutter any of the pages with requires
or includes to bootstrap.