jp - add codeclimate coverage metric to project
Also updated the .gitignore so that when you push the project after running
coverage tests locally, you won't be adding clutter to the team repo. Also added
a new raketask that was copied from the piazza thread that Dorthy created and
updated the README.md to actually contain the button that shows our coverage.